requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
aioredis==1.3.1
asgiref==3.4.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.3
boto3==1.18.49
botocore==1.21.49
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dateparser==1.0.0
dj-database-url==0.5.0
Django==3.2.6
django-background-tasks==1.2.5
django-bootstrap-v5==1.0.5
django-compat==1.0.15
django-cors-headers==3.8.0
django-mailjet==0.3.1
django-ses==2.3.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.39.0
gunicorn==20.1.0
h5py==3.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
jmespath==0.10.0
keras==2.6.0
Keras-Preprocessing==1.1.2
mailjet-rest==1.3.4
Markdown==3.3.4
msgpack==1.0.2
multidict==5.1.0
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.2
protobuf==3.17.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-binance==1.0.12
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
service-identity==21.1.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
ta==0.7.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.1.0
urllib3==1.26.6
websockets==9.1
Werkzeug==2.0.1
whitenoise==5.3.0
wrapt==1.12.1
yarl==1.6.3
zope.interface==5.4.0
